+++ playa/font.py
@@ -1,6 +1,7 @@
 import logging
 from io import BytesIO
 from typing import (
+    cast,
     Dict,
     Iterable,
     List,