+++ tests/test_fonts.py
@@ -40,8 +40,8 @@ Un peu plus à droite"""
         )
         # Did we get the *glyphs* right? (harder)
         boxes = list(t.bbox for t in page.texts)
-        assert boxes[0] == pytest.approx((100.0, 72.968, 289.408, 96.968))
-        assert boxes[1] == pytest.approx((150.0, 108.968, 364.776, 132.968))
+        assert boxes[0] == pytest.approx((100.0, 74.768, 289.408, 96.968))
+        assert boxes[1] == pytest.approx((150.0, 110.768, 364.776, 132.968))
 
 
 @pytest.mark.skipif(not CONTRIB.exists(), reason="contrib samples not present")
@@ -89,7 +89,7 @@ def test_type3_font_boxes() -> None:
         textor = page.texts
         line1 = next(textor).bbox
         assert line1 == pytest.approx(
-            (25.0, 14.3701175326, 246.586936753, 28.3701175326)
+            (25.0, 14.274413, 246.586937, 28.370118)
         )
         boxes: List[Rect] = []
         for text in textor:
@@ -102,5 +102,5 @@ def test_type3_font_boxes() -> None:
             boxes.append(bbox)
         line2 = get_bound_rects(boxes)
         assert line2 == pytest.approx(
-            (25.0, 39.3701175326, 246.58691507160006, 53.3701175326)
+            (25.0, 39.274413, 246.58691507160006, 53.3701175326)
         )